[GEO] Packed QuadTree
Lucene's QuadPrefixTree creates a tremendous amount of string terms for each quadcell. This can be optimized by representing these cells as a packed binary Long value. These values are then streamed to the DocumentWriter instead of passing an Array iterator which cuts memory usage in half. This enhancement does not fix the quad tree memory problem. It defers it for large funky shapes. More optimizations coming.